requirements.txt
riak==2.0.3
pillow==2.4.0
python-magic==0.4.6
